+++ src/NotFoundException.php
@@ -6,6 +6,7 @@ namespace Yiisoft\Di;
 
 use Exception;
 use Psr\Container\NotFoundExceptionInterface;
+use Throwable;
 
 /**
  * `NotFoundException` is thrown when no definition or class was found in the container for a given ID.
@@ -18,19 +19,34 @@ final class NotFoundException extends Exception implements NotFoundExceptionInte
      */
     public function __construct(
         private string $id,
-        array $buildStack = []
+        private array $buildStack = [],
+        ?Throwable $previous = null,
     ) {
-        $message = $id;
-        if ($buildStack !== []) {
-            $last = end($buildStack);
-            $message = sprintf('%s" while building %s', $last, '"' . implode('" -> "', $buildStack));
+        if (empty($this->buildStack)) {
+            $message = sprintf('No definition or class found or resolvable for "%s".', $id);
+        } elseif ($this->buildStack === [$id]) {
+            $message = sprintf('No definition or class found or resolvable for "%s" while building it.', $id);
+        } else {
+            $message = sprintf(
+                'No definition or class found or resolvable for "%s" while building "%s".',
+                end($this->buildStack),
+                implode('" -> "', $buildStack),
+            );
         }
 
-        parent::__construct(sprintf('No definition or class found or resolvable for "%s".', $message));
+        parent::__construct($message, previous: $previous);
     }
 
     public function getId(): string
     {
         return $this->id;
+    }
+
+    /**
+     * @return string[]
+     */
+    public function getBuildStack(): array
+    {
+        return $this->buildStack;
     }
 }